backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
pdfplumber==0.11.7
python-docx==1.2.0
transformers==4.53.2
torch==2.4.0
sentence-transformers==5.0.0
faiss-cpu==1.11.0.post1
numpy==2.2.6
rank-bm25==0.2.2
requests==2.31.0
